move scanPage to own directory scan
